Models/Question Generation/requirements.txt
google-generativeai==0.8.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
streamlit==1.39.0
streamlit-option-menu==0.4.0
plotly==5.23.0
aiohttp==3.10.10
